requirements.txt
# fastapi==0.115.0
# uvicorn[standard]==0.30.6
# sqlalchemy==2.0.36
# asyncpg==0.30.0
# alembic==1.13.2

# python-dotenv==1.0.1

# passlib[bcrypt]==1.7.4
# bcrypt==4.0.1

# python-jose==3.3.0
jinja2==3.1.4






fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
emails==0.6
python-dateutil==2.8.2
pytz==2023.3.post1
bcrypt==4.0.1
